+++ packages/provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.28.7"></a>
+## [0.28.7](https://github.com/newsuk/times-components/compare/@times-components/provider@0.28.6...@times-components/provider@0.28.7) (2018-05-03)
+
+
+
+
+**Note:** Version bump only for package @times-components/provider
+
 <a name="0.28.6"></a>
 ## [0.28.6](https://github.com/newsuk/times-components/compare/@times-components/provider@0.28.5...@times-components/provider@0.28.6) (2018-04-30)